+++ crates/holochain/src/core/workflow/sys_validation_workflow.rs
@@ -12,7 +12,6 @@ use holochain_cascade::Cascade;
 use holochain_cascade::CascadeImpl;
 use holochain_p2p::HolochainP2pDnaT;
 use holochain_services::DpkiMutex;
-use holochain_services::DpkiService;
 use holochain_sqlite::prelude::*;
 use holochain_state::prelude::*;
 use parking_lot::Mutex;
@@ -422,7 +421,7 @@ pub(crate) async fn validate_op(
     op: &DhtOp,
     dna_def: &DnaDefHashed,
     validation_dependencies: Arc<Mutex<ValidationDependencies>>,
-    dpki: DpkiMutex,
+    dpki: Option<DpkiMutex>,
 ) -> WorkflowResult<Outcome> {
     match validate_op_inner(op, dna_def, validation_dependencies, dpki).await {
         Ok(_) => Ok(Outcome::Accepted),
@@ -503,7 +502,7 @@ async fn validate_op_inner(
     op: &DhtOp,
     dna_def: &DnaDefHashed,
     validation_dependencies: Arc<Mutex<ValidationDependencies>>,
-    dpki: DpkiMutex,
+    dpki: Option<DpkiMutex>,
 ) -> SysValidationResult<()> {
     check_entry_visibility(op)?;
     if let Some(dpki) = dpki {
@@ -903,7 +902,7 @@ pub struct SysValidationWorkspace {
     cache: DbWrite<DbKindCache>,
     pub(crate) dna_def: Arc<DnaDef>,
     sys_validation_retry_delay: Duration,
-    dpki: DpkiMutex,
+    dpki: Option<DpkiMutex>,
 }
 
 impl SysValidationWorkspace {
@@ -913,7 +912,7 @@ impl SysValidationWorkspace {
         dht_query_cache: DhtDbQueryCache,
         cache: DbWrite<DbKindCache>,
         dna_def: Arc<DnaDef>,
-        dpki: DpkiMutex,
+        dpki: Option<DpkiMutex>,
         sys_validation_retry_delay: Duration,
     ) -> Self {
         Self {
